+++ src/index.ts
@@ -1,5 +1,4 @@
 export * from './controllers';
-export * from './constants';
 export * from './services';
 export * from './types';
 export * from './utils';