+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Removed constants index file](https://github.com/multiversx/mx-sdk-dapp/pull/1674)
 - [Added LogoutManager tests](https://github.com/multiversx/mx-sdk-dapp/pull/1673)
 - [Fixed issue with config changes for nativeAuth and walletConnect](https://github.com/multiversx/mx-sdk-dapp/pull/1672)